Print age of persons' birthdays
When a line is found having the text "geboren XXXX", where XXXX should be the year of the person's birth, the age of the person is calculated and printed.

Example line of daterem.dat:
25.10. Test-Geburtstag, geboren 1981

The output on 2011-10-25 is:
Di, 25.10.2011 Test-Geburtstag, geboren 1981, Alter: 30 Jahre

The english word "born" can be used too. So "born 1981" is also captured. There has to be a space or a comma or something like that (regex \W) in front of geboren/born in order to work correctly.